Split debuginfo/emit.rs part out of debuginfo/mod.rs
